Improved Layout types
Use enums that precisely model what data is stored.
Follows what was written in the Readme.

These structures take into account serde-friendlyness:
- ordered structures instead of HashMap
- ordering means we can use indexes for relations+primary.

Relations should be stored in a linearized triangular matrix.
This is WIP, but is better as there are no redundant values,
like in a full matrix, nor badly defined sparsity in the case of hashmap.

Added comments to improve cargo doc.
Removed caret from dependency version as it is not needed.