query_output_state: hashmaps, more checking
Used hashmaps output_id -> output_state (and same for crtc).
This will be needed as ids are not in sequence and manipulating these
objects requires a random access pattern.
It is kind of overkill to use hashmap, but this is semantically correct.

Added check for status field in some requests.
For now generate a hard error, as in the python version they never failed.

Lots of style improvements.