Put layout inference in layout module
Computing manually the rects then the relations between them in mod:xcb
would require a way to translate output ids to the order in Layout
sorted vec.

A simpler way is to give a vec<(state,rect)> to layout::new and let it
sort internally both data in sync, according to state.id().
This way we can "unzip" the vec into one vec with states and one with
rects, same ordering, and use the ones with vec for relation map
inference.

This requires some complexity to generate the (state, option<rect>)...