Move to "normal" lib+bin configuration
I previously avoided doing a lib+bin, as it prevents internal
documentation to be generated for the bin-only code. Only the lib part
would have documentation.

However this meant reusing files from the bin in a hackish way for the
example. This was okay, but would have broken with the new nested
compute_rects module.

Thus move to proper lib+bin. As long as we keep the default bin name,
just adding a main.rs and a lib.rs works. Documentation is ok as long as
the bin code is kept minimal (only clap and overall startup). The lib
part makes no sense in isolation. The compute rect code and private
structs have been moved into their own file.

Overall organisation between lib & bin will be revamped as the bin code
is finalized (with db & stuff).