Remove simplification of singleton constraints
Singleton constraints can be simplified to constant (mono case) or equalities (dual case).
However this introduces lot of complexity, for something that may not happen often.

The complexity is due to chained simplifications ; for example :
- a replace_by_const transforms a dual constraint into mono
- the mono constraint as singleton triggers another replace by const
In this case it becomes difficult to organize simplifications without
breaking stored indexes and other temporary data.

The main goal of the QpProblem pre-processing is to reduce the number of
variables given to the optimizer. 50% of our constraints are equalities
which means pre-processing of these is effective, and used all the time.
Singleton constraints is a "nice to have" but may not trigger as often,
so it is ok to drop for the moment. Maybe it can be reintroduced as a
later step.