Model shift : store base coords instead
Using autolayout all the time is wasteful, and makes specialized layouts
(bottom aligned, etc) difficult to represent.
In addition it must be reliable to ensure it can always generate a
layout.
Another problem is complexity of the code, with the need to "thread" the
preferred modes everywhere, etc...

- moved all direction and auto layout stuff in unused files for the time being
- changed layout to use coordinates instead of relations

The goal is to still have an "auto layout" function, but only use it
when a new layout must be generated, or to normalise on demande.