Restructure status
status must not be stored in the database, so move it away from layout content.
Instead return it bundled with the layout, easy to separate.
This also allows xcb backend to post process it for custom checks.

Start implementing the reaction loop.
For now mostly ignored cases with log message, and TODO placeholder.
Next step : database, then apply layout.