Make OutputEntry Ord
Sorting OutputEntry with only ids may be non deterministic if an Edid
appears multiple times, for example if a screen is connected using two
cables and inputs.

A solution is to sort OutputEntry on all fields.
Thus Ord derives have been added to many structs.

The f64 frequency field in Mode has been replaced with an integer
frequency to make Ord simple (as f64 is not Ord).
The solution is to round the value, which is deterministic in the xcb
backend at least.