Apply stored layout from database
Add daemon logic code for "apply from database".

Sadly it was not possible to use the alternate key mode in hashset to only match with something representing the list of output ids.
This is due to Borrow trait which is very constrained : only allows a reference to existing object as alternate key.
By making LayoutById public, we can temporarily build one for matching, and recycle the inner layout if needed later.

Added "apply_layout" method to the backend trait, this is the next thing to do.
Afterwards it will be autolayouting (ouch !).