Unsupported case semantics for duplicate EDID
Duplicate EDID are allowed as they can happen : multiple connections to the monitor, this happens in my current setup.
However, duplicate ENABLED edid are unsupported as no backend will support them.
Add an unsupported check for that (annoying to do without groupby).

In addition, we want to load a stored layout in cas we encounter duplicate enabled EDID, for example ubuntu stupid autolayout.
Thus changed unsupported handling : only ignore them if they were to be stored in the DB.