Complete precomputing of association 'xcb_output' to config
Now that we do not need to support duplicated EDIDs, this function is easy.

Previous iterations with duplicate EDID support required to generate a
list of possible output use, and try to sort them by some criterion.
The criterion was by mode "capability" of outputs, basically taking the
"best" mode supported by one output as measure of capability.