Remove clone unsupported cause
Experiments on a laptop showed that some outputs, even unused, were declared with active clones permanently.
This seems to be a way to indicate the equivalence of DP with its virtual HDMI output in DP altmode on this particular machine.

The problem is that even if the layout was not using these, it would be tagged as unsupported and never stored.
The check is in fact redundant as we already check for coordinate overlap, which is the real "bad" case.
Thus this check has been dropped entirely.

Keeping the overlap check is still useful, as except for the clone case, any overlap would break many window managers.
And with an overlap it would be difficult to do autolayouting.