Readme update
Tried edid_rs for better parsing of Edid.
This can extract monitor names (when defined), subfields of EDID, etc.
It could be used later to identify monitors by serial number + manufacturer, and recognize it even if behind another output. Not a priority though.
On the Dell work laptop, the internal display has no name, so this is not that useful even for debugging.