Partial apply_layout impl
Rewritten error management of this part.
This requires explicit split between recoverable errors and fatal errors, but both use Result to have proper logging.
The only use of panic is for a few overflows that are very unlikely.

The rust system uses the same sequence of set_crtc as the python one.
Small tests on the laptop shows some problems, hard to say if this is
due to the gnome layouting daemon, or DP over usb-C problems.
Maybe due to dynamic outputs not properly disconnecting ?